Add generic `Gpio` type
This type is useful to implement other peripherals that need access to
GPIO pins.

I've added a small TODO to implement into_*_output function for this
type, I'm not sure how to proceed since the esp32 with
`funcX_out_sel_cfg` is unergonomic and the code would need to match over
each pin to use the correct register. This can be resolved by making
these registers an array.

Signed-off-by: Jean Pierre Dudey <jeandudey@hotmail.com>